Breakup solution into seperate files
